requirements.txt
langchain_core
python_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
pypdf
faiss-cpu
sentence_transformers
langchain
huggingface-hub
dateparser

